displayer invitation done & ready to be tested
